Fix: Fix the Add contribution functionalities to update progress on the Goal Table
