+++ src/Context/GoalsContext.jsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useState, useEffect } from 'react';
+import React, { createContext, useContext, useState, useEffect } from "react";
 
 export const GoalsContext = createContext();
 
@@ -16,15 +16,37 @@ export const GoalsProvider = ({ children }) => {
   };
 
   const toggleComplete = (id) => {
-    const updatedGoals = goals.map(goal =>
+    const updatedGoals = goals.map((goal) =>
       goal.id === id ? { ...goal, completed: !goal.completed } : goal
     );
     setGoals(updatedGoals);
     localStorage.setItem("goals", JSON.stringify(updatedGoals));
   };
 
+  const addContribution = (goalId, contributionAmt) => {
+    const updatedGoals = goals.map((goal) => {
+      if (goal.id === Number(goalId)) {
+        return {
+          ...goal,
+          currentAmt: Number(goal.currentAmt || 0) + Number(contributionAmt),
+        };
+      }
+      return goal;
+    });
+    setGoals(updatedGoals);
+    localStorage.setItem("goals", JSON.stringify(updatedGoals));
+  };
+
+  const deleteGoal = (goalId) => {
+    const updatedGoals = goals.filter((goal) => goal.id !== goalId);
+    setGoals(updatedGoals);
+    localStorage.setItem("goals", JSON.stringify(updatedGoals));
+  };
+
   return (
-    <GoalsContext.Provider value={{ goals, createGoal, toggleComplete }}>
+    <GoalsContext.Provider
+      value={{ goals, createGoal, toggleComplete, addContribution, deleteGoal }}
+    >
       {children}
     </GoalsContext.Provider>
   );
